+++ src/store/models/Activity.js
@@ -1,10 +1,10 @@
 import { attr, fk, Model } from 'redux-orm'
 
-const Activity = Model.createClass({
+class Activity extends Model {
   toString () {
     return `Message: ${this.id}`
   }
-})
+}
 
 export default Activity