+++ src/store/selectors/getTopicForCurrentRoute.js
@@ -5,7 +5,6 @@ import presentTopic from 'store/presenters/presentTopic'
 
 const getTopicForCurrentRoute = ormCreateSelector(
   orm,
-  state => state.orm,
   (state, props) => getRouteParam('topicName', state, props),
   (session, topicName) => {
     const topic = session.Topic.safeGet({ name: topicName })